Deletion logic works to delete the corresponding uploads on the file system

+++ controllers/admin.js
@@ -1,3 +1,4 @@
+const fileHelper = require('../util/file');
 const { validationResult } = require('express-validator');
 const Product = require('../models/product');
 
@@ -168,8 +169,9 @@ exports.postEditProduct = (req, res, next) => {
       product.description = updatedDesc;
 
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
-      }
+      };
 
       return product.save().then((result) => {
         console.log('UPDATED PRODUCT');
@@ -202,7 +204,16 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('Product not found'));
+      };
+
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id })
+    })
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
